# apps/backend/internal/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/links/pairing-code": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Links"
                ],
                "summary": "Create a short-lived patient pairing code",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/server.PairingCodeCreateResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/links/pairing-code/redeem": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Links"
                ],
                "summary": "Redeem a patient pairing code",
                "parameters": [
                    {
                        "description": "Redeem payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.PairingCodeRedeemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.LinkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "server.LinkResponse": {
            "type": "object",
            "properties": {
                "link": {
                    "$ref": "#/definitions/server.linkDTO"
                },
                "patient": {
                    "$ref": "#/definitions/server.patientDTO"
                }
            }
        },
        "server.PairingCodeCreateResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "string"
                },
                "qrText": {
                    "type": "string"
                }
            }
        },
        "server.PairingCodeRedeemRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
...
        "server.linkDTO": {
            "type": "object",
            "properties": {
                "approvedAt": {
                    "type": "string"
                },
                "doctorId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "patientId": {
                    "type": "string"
                },
                "requestedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "server.patientDTO": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "patientCode": {
                    "type": "string"
                }
            }
        },
...
    }
}